+++ src/main/java/frc/lib5k/components/motors/MixedMotorCollection.java
@@ -9,7 +9,7 @@ import edu.wpi.first.wpilibj.SpeedControllerGroup;
 import frc.lib5k.components.motors.interfaces.IMotorCollection;
 import frc.lib5k.components.motors.interfaces.IRampRateController;
 import frc.lib5k.components.motors.interfaces.IVoltageOutputController;
-import frc.lib5k.control.TimedSlewLimiter;
+import frc.lib5k.control.motionprofile.TimedSlewLimiter;
 import frc.lib5k.interfaces.Loggable;
 import frc.lib5k.roborio.RR_HAL;
 import frc.lib5k.utils.ObjectCounter;
